# core/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Archura Core API",
    "version": "0.1.0",
    "description": "Identity, ownership, and component-session API for Archura. Production /v1 requests require the Worker service credential plus any endpoint-specific bearer credential."
  },
  "servers": [
    { "url": "/", "description": "Current core server" }
  ],
...
  "components": {
    "securitySchemes": {
      "ServiceAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Archura-Service-Authorization",
        "description": "Worker credential including the Bearer prefix, for example `Bearer svc_live_...`. Required in production and optional in direct local development."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "adm_..., sk_..., or sess_...",
        "description": "Platform-admin key, organization secret key, or account session token, according to the operation."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "invalid_api_key" },
              "message": { "type": "string", "example": "The organization API key is invalid." },
              "request_id": { "type": "string", "example": "core/abc123-000001" }
            }
          }
        }
      },
...
      "CreateComponentSessionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["component_id", "origin"],
        "properties": {
          "component_id": { "type": "string", "pattern": "^cmp_(?:test|live)_", "example": "cmp_test_..." },
          "external_user_id": { "type": "string", "description": "Optional organization-owned user reference. Never written to operational or audit logs." },
          "origin": { "type": "string", "format": "uri", "example": "https://mikesbakery.example" }
        }
      },
      "ComponentSessionCreated": {
        "type": "object",
        "required": ["access_token", "token_type", "expires_in", "expires_at"],
        "properties": {
          "access_token": { "type": "string", "pattern": "^ct_(?:test|live)_", "example": "ct_test_...", "description": "Short-lived browser bearer token. Returned once and stored only as a hash." },
          "token_type": { "type": "string", "const": "Bearer" },
          "expires_in": { "type": "integer", "const": 600 },
          "expires_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Organization": {
        "type": "object",
        "required": ["id", "name", "slug", "role", "is_default", "publishable_key", "sites", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "role": { "type": "string", "enum": ["owner", "member"] },
          "is_default": { "type": "boolean" },
          "publishable_key": { "type": "string", "pattern": "^pk_(?:test|live)_" },
          "sites": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateOrganizationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "slug": { "type": "string", "pattern": "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$" },
          "allowed_origins": { "type": "array", "items": { "type": "string", "format": "uri" } }
        }
      },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid service/admin/organization credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Authenticated but not allowed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "Resource conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "RateLimited": { "description": "Rate limit exceeded", "headers": { "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "InternalError": { "description": "Internal error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unavailable": { "description": "Database unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  },
  "paths": {
...
    "/v1/component-sessions": {
      "post": {
        "tags": ["Sessions"],
        "operationId": "createComponentSession",
        "summary": "Mint a short-lived component bearer token",
        "description": "Called by the organization backend after it authenticates its own user. The token is bound to the organization, component, checkout:create scope, checkout audience, origin, and ten-minute expiry.",
        "security": [{ "ServiceAuth": [], "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateComponentSessionRequest" } } } },
        "responses": {
          "201": { "description": "Component session created", "headers": { "Cache-Control": { "schema": { "type": "string", "example": "no-store" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComponentSessionCreated" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
...
    "/v1/organizations": {
      "post": {
        "tags": ["Organizations", "Accounts"],
        "operationId": "createOrganization",
        "summary": "Create another organization for the current account",
        "security": [{ "ServiceAuth": [], "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrganizationRequest" } } } },
        "responses": {
          "201": { "description": "Organization created; secret_key is returned once", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Organization" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
...
  }
}